move getFormField to HttpUtil.
